% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exVCFdoc.R
\docType{data}
\name{exVCF}
\alias{exVCF}
\title{This is example VCF data}
\format{An object of class \code{CollapsedVCF} with 5 rows and 6 columns.}
\usage{
exVCF
}
\description{
This is example VCF data
}
\author{
RMHamner
}
\keyword{data}
