% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2impostar.R
\name{vcf2impostar}
\alias{vcf2impostar}
\title{Convert VCF to impostar format}
\usage{
vcf2impostar(vcfFile="filename", designFile=NULL, ploidy=2, savecsv=FALSE)
}
\arguments{
\item{vcfFile}{a character string indicating the vcf file you want to convert to impostar format. Required.}

\item{designFile}{a character string indicating the design file, which contains the number of indiviudals in each pool. Only required for pools. Default is NULL and should be left this way for individuals. The file can be .csv or .txt (tab-delimited)}

\item{ploidy}{a numeric indicating the ploidy of the focal organism. Default is 2. Note for pooled datasets: ploidy should be set to the ploidy of a single organism, the number of alleles included in a pool should be indicated in the designFile.}

\item{savecsv}{a logical indicating if you want to save the output as a csv file. If FALSE (default), the output will only be stored as an R object.}
}
\value{
The vcf2impostar function returns a dataframe and optional csv file written to the working directory. The dataframe contains a row for each SNP locus and columns containing a unique name for each SNP based on its sequence location (snpID), reference allelic state (REF), alternate allelic state (ALT), reference reads (RD), alternate reads (AD), total read depth (DP), and total number of alleles (2n, TotAlleles) per individual or pool (as indicated by sequential numbers in the column names).
}
\description{
This function reformats a VCF file into the format required by impostar functions runLogRegTest and runAMOVA
}
\examples{
# load example vcf file
myvcfFile <- impostar::: exVCF
# generate an example designFile
designFile <- data.frame(n=rep(20,6), Sample=c(seq(1:6)), Region=c(1,1,2,2,3,3))
# run vcf2impostar format converter
data.impostar <- vcf2impostar(vcfFile = vcfFile, designFile = designFile, ploidy = 2, savecsv = FALSE)
# to feed it directly into runAMOVA
testAMOVA <- runAMOVA(designFile = designFile, dataFile = data.impostar, NresamplesToStop = 5, maxPermutations = 10)
}
\references{
Hamner, R.M., J.D. Selwyn, E. Krell, S.A. King, and C.E. Bird. In review. Modeling next-generation sequencer sampling error in pooled population samples dramatically reduces false positives in genetic structure tests.
}
\seealso{
\code{\link{runLogRegTest}}, \code{\link{runAMOVA}}
}
\author{
Rebecca M. Hamner
}
